requirements.txt
asgiref==3.7.2
Django
djangorestframework==3.14.0
psycopg2-binary
python-dotenv
pytz==2023.3
sqlparse==0.4.4
